Fixed a return type into @GeometricConsistencyGrouping@ Added unit tests for @Hough3DGrouping@ and @GeometricConsistencyGrouping@
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@5302 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ recognition/include/pcl/recognition/cg/geometric_consistency.h
@@ -82,7 +82,7 @@ namespace pcl
         * 
         * \return the minimum cluster size used by GC.
         */
-      inline double
+      inline int
       getGCThreshold () const
       {
         return (gc_threshold_);
@@ -123,7 +123,7 @@ namespace pcl
       using CorrespondenceGrouping<PointModelT, PointSceneT>::model_scene_corrs_;
 
       /** \brief Minimum cluster size. It shouldn't be less than 3, since at least 3 correspondences are needed to compute the 6DOF pose */
-      double gc_threshold_;
+      int gc_threshold_;
 
       /** \brief Resolution of the consensus set used to cluster correspondences together*/
       double gc_size_;
